requirements.txt
diffusers==0.26.3
Pillow==10.0.1
PySide6==6.6.2
torch==2.2.0
xformers==0.0.24
accelerate==0.27.0
transformers==4.37.2
invoke
